use form

+++ database/migrations/2023_09_04_013815_create_listings_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedTinyInteger('beds');
             $table->unsignedTinyInteger('baths');
-            $table->unsignedTinyInteger('area');
+            $table->unsignedInteger('area');
             $table->tinyText('city');
             $table->tinyText('code');
             $table->tinyText('street');
